requirements.txt
asgiref==3.8.1
asttokens==3.0.0
attrs==24.3.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.2.0
certifi==2024.12.14
charset-normalizer==3.4.1
colorama==0.4.6
contourpy==1.3.1
cycler==0.11.0
decorator==5.1.1
dedup==0.0.2
defusedxml==0.7.1
Django==5.1.4
docopt==0.6.2
executing==2.1.0
fastjsonschema==2.21.1
filelock==3.13.1
fire==0.7.0
fsspec==2024.2.0
idna==3.10
image==1.5.33
ImageHash==4.3.1
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.3
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.4
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mistune==3.1.0
mkl-service==2.4.0
mkl_fft==1.3.11
mkl_random==1.2.8
mpmath==1.3.0
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
networkx==3.2.1
numpy==1.26.4
opencv-python==4.10.0
opencv-python-headless==4.10.0
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pillow==11.0.0
pipreqs==0.5.0
platformdirs==4.3.6
ply==3.11
prompt_toolkit==3.0.48
pure_eval==0.2.3
Pygments==2.19.1
pyparsing==3.2.0
PyQt5==5.15.10
PyQt5-sip==12.13.0
PyWavelets==1.7.0
pywin32==308
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rpds-py==0.22.3
scikit-learn==1.6.0
scipy==1.14.1
sip==6.7.12
six==1.16.0
soupsieve==2.6
sqlparse==0.5.3
stack-data==0.6.3
sympy==1.13.1
termcolor==2.5.0
threadpoolctl==3.5.0
tinycss2==1.4.0
tomli==2.0.1
torch==2.5.1+cu118
torchaudio==2.5.1+cu118
torchvision==0.20.1+cu118
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.9.0
tzdata==2024.2
unicodedata2==15.1.0
urllib3==2.3.0
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
